requirements.txt
# LangChain core and Google GenAI
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.7

# Google Generative AI
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6

# PDF handling
pypdf==3.11.0

# Transformers dependency required by some LangChain components
transformers==4.57.3
tokenizers==0.22.1

# Optional: dotenv for environment variables
python-dotenv==1.0.1

# Vector stores
chromadb==0.5.0
faiss-cpu==1.7.4

streamlit
